+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui-dev
 
+## 1.0.7
+
+### Patch Changes
+
+- update sapa hook document
+
 ## 1.0.6
 
 ### Patch Changes